+++ ZendeskSell/Collaborations/ICollaborationActions.cs
@@ -3,7 +3,7 @@ using ZendeskSell.Models;
 
 namespace ZendeskSell.Collaborations {
     public interface ICollaborationActions {
-        Task<ZendeskSellCollectionResponse<CollaborationResponse>> GetAsync(int pageNumber, int numPerPage, int? creatorID = null, int? resourceID = null, string resourceType = null);
+        Task<ZendeskSellCollectionResponse<CollaborationResponse>> GetAsync(int pageNumber, int numPerPage, int? creatorID = null, long? resourceID = null, string resourceType = null);
         Task<ZendeskSellObjectResponse<CollaborationResponse>> GetOneAsync(int id);
         Task<ZendeskSellObjectResponse<CollaborationResponse>> CreateAsync(CollaborationRequest collaboration);
         Task<ZendeskSellDeleteResponse> DeleteAsync(int id);
